DASB: Tokotron: Refactoring for standalone evaluation
